Module pallet_price_oracle::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_default_parts_6 as tt_default_parts;

Structs

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The pallet implementing the on-chain logic.

_GeneratedPrefixForStorageNewAssets

_GeneratedPrefixForStoragePriceDataAccumulator

_GeneratedPrefixForStoragePriceDataHundred

_GeneratedPrefixForStoragePriceDataTen

_GeneratedPrefixForStoragePriceDataThousand

_GeneratedPrefixForStorageTrackedAssetsCount

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Type Definitions

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

NewAssets

Sorted array of newly registered assets. All assets are processed and removed from the storage at the end of a block. Trades start to be processed from the next block. All trades in the same block as the asset registration are ignored.

PriceDataAccumulator

Processed or partially processed data generated by trades. Data generated by trades are processed sequentially. Each new entry is combined with the previous value to produce new intermediate value. The last entry creates the resulting average price and volume.

PriceDataHundred

The last ten average values corresponding to the last hundred blocks. Each average value corresponds to an interval of length ten blocks.

PriceDataTen

The last ten average values corresponding to the last ten blocks.

PriceDataThousand

The last ten average values corresponding to the last thousand blocks. Each average value corresponds to an interval of length hundred blocks.

TrackedAssetsCount

The number of assets registered and handled by this pallet.