Type Definition pallet_price_oracle::pallet::PriceDataAccumulator

pub type PriceDataAccumulator<T: Config> = StorageMap<_GeneratedPrefixForStoragePriceDataAccumulator<T>, Twox64Concat, AssetPairId, PriceEntry, ValueQuery>;

Processed or partially processed data generated by trades. Data generated by trades are processed sequentially. Each new entry is combined with the previous value to produce new intermediate value. The last entry creates the resulting average price and volume.